initial-commit
This is the first version of this library. I have been inspired by one of my studnets #shanitah to develop this simple library it is simple and easy to understant yet powerfull.

Since this is the initail write. It is meant for beginners who would like to have powerfull web applications with view components inside thier  javascript and also have the ablity to setup state in thiser applications.